Move debounce utility from CLI to core workspace
The debounce function was located in src/cli/src/shared/debounce.ts, which
is an inappropriate location for a general-purpose function composition utility.

Original location: src/cli/src/shared/debounce.ts
New location: src/core/src/utils/function.ts

Rationale:
- The debounce function is a pure, generic timing/function composition utility
  with no CLI-specific behavior or dependencies
- According to AGENTS.md: "Code must be organized by domain, not by generic
  utility patterns"
- The core workspace already has a utils/function.ts module that houses other
  function utilities like identity(), noop(), and callWithFallback()
- Moving debounce to function.ts groups it with conceptually similar utilities
- This makes debounce available to all workspaces (parser, plugin, transpiler,
  etc.) rather than being artificially limited to CLI code

Changes:
- Added debounce function and DebouncedFunction interface to core/utils/function.ts
- Moved test from cli/test/shared-debounce.test.ts to core/test/function-debounce.test.ts
- Updated imports in cli/commands/watch.ts to use Core.debounce
- Removed debounce.ts from CLI shared directory
- Exported DebouncedFunction type from core package for external use
- All tests pass, build succeeds
